Populate an old Chrome CVE.
This was allocated by Chrome, possibly erroneously, on the basis
that Linux was an open source component of ChromeOS and is therefore
in our CNA scope.

However this bug never affected a released version of ChromeOS.

As this is now publicly referenced in various places, it probably
makes most sense just to populate it.