Put CVE-2018-16301 right
This CVE was allocated on 2018-09-01 by MITRE via request scr562827 for
a vulnerability in tcpdump.

Fixing the vulnerability took much longer than expected.  Meanwhile some
erroneous updates in tcpdump and libpcap change logs resulted in a
number of erroneous updates in the CVE in October-December 2019.  In
March 2020 the CVE was marked as rejected because of the confusion.

The original vulnerability was eventually fixed in September 2020 and
tcpdump release 4.99.0 in December 2020 was the first to include the
fix.  Reinstate the CVE as "public" and add information from the
original CVE assignment request as well as subsequent developments.

See also https://github.com/the-tcpdump-group/libpcap/issues/855