Reject CVE-2016-1000027
As explained in spring-projects/spring-framework#24434, this CVE has
been reported through a now disbanded CNA and is invalid.

A similar issue has been reported with CVE-2011-2894, which got closed
with documentation updates. This is not a vulnerability in Spring
Framework, but rather a well-known security concern in all Java
applications.

This commit rejects this CVE as a duplicate of CVE-2011-2894 as a
result.