Update CVE-2022-40144.json
I found that the character encoding rule used by the CVE-2022-40144.json file is not standard utf-8. When I use a program to parse the json file, errors will occur, while other files will not. The encoding method should be the same according to the CVE ID JSON File Format, So I modified the exception code of the file to make it look better.